Pre-emptively open a session to the PKCS#11 base slot.
Before this change, every aziot-keys function would open a new session, log in,
then close the session. For operations like getting a public key, there would
be many calls to get each of the public key's parameters.

For some PKCS#11 libraries like tpm2-pkcs11, logging in is very slow (~5s).
This would make such operations very slow.

With this change, the aziot-keys implementation pre-emptively opens a session
to the base slot when both the lib path and the base slot are set on it.
The underlying PKCS#11 wrapper caches open sessions, so by holding on to this
pre-emptively opened session future requests automatically reuse it instead of
opening new ones.