Implement derived keys.
Derived keys are a third kind of key handled by libaziot-keys and aziot-keyd.
The derivation uses a base key and arbitrary data. libaziot-keys combines these
in some implementation-defined way to create a new symmetric key, that can then
be used for signing, encryption and decryption.

Derived keys can also be exported to plaintext. This is required, for example,
for IS to register module SAS keys with IoT Hub. This sort of makes
the sign / encrypt / decrypt support for derived keys via
the mechanism redundant, since the caller could export the key and perform
the operation itself. But it's possible that this API may be restricted to
privileged callers in the future, like IS as mentioned above, while
the sign / encrypt / decrypt APIs would be more freely accessible.

Other changes:

- Make key filenames incorporate the ID just like cert filenames.