Avoid unnecessary IoT Hub call to list modules (#37)
In `identity::Server::init_identities(...)`, we query IoT Hub for a list of modules to decide whether any identities should be added or removed. But if the function arguments (`prev_module_set` and `current_module_set`) are both empty, there's no need to make the cloud API call because the function is effectively a no-op.

This change adds an early-exit check to the function, to avoid the cloud API call when it's not necessary.