Disallow custom endpoints and HTTP servers in release builds.
They can still be used with debug builds for testing, but release builds
must use the standard endpoints, and must use UDS or systemd endpoints.
Even if they have been set in the service config file, release builds will
ignore them.

Because the endpoints are not configurable, this means the servers can
no longer use *either* systemd socket activation *or* bind a socket themselves.
They must do the former and *fall back* to the latter if the systemd env vars
haven't been set. This is how socket-activated services should work anyway,
so this only makes our services consistent with other software. As a result,
the fd scheme now also contains a query string parameter that specifies
the fallback URI.

This commit also standardizes the sockets directory to be `/run/aziot`
rather than the mix of `/var/lib/aziot` on distros that did not use
systemd socket activation (CentOS 7) and `/var/run/aziot` on distros that did
(everything else). This is needed to ensure that arbitrary host modules always
know the path of the unix socket for the service they want to connect to,
instead of needing it to be configured by the device operator in some way.
Unlike with iotedged, there is no env var that could give this information to
an arbitrary host module, because host modules run independently of
our services. Furthermore, all our supported distros already have `/run`,
so there's no reason to continue using `/var/run`.

Next, this commit enables systemd socket activation on CentOS 7. The original
decision to disable socket activation on CentOS 7 was inherited from iotedged,
which does it because CentOS 7 doesn't support `LISTEN_FDNAMES` and thus
cannot be used for socket-activated services that want more than one socket.
This is the case for iotedged with its mgmt and workload sockets, but is not
the case for the services in this repo. So the `fd://` endpoint URIs no longer
have a hostname that identifies a particular fd by number or name, and instead
the services always take the first fd they get as long as `LISTEN_FDS` is
greater than 0.

Thus, putting all these changes together, the default endpoint URIs are
`unix:///run/aziot/<service_name>.sock` URIs for other services' endpoints
and
a `fd://?fallback=urlencode(unix:///run/aziot/<service_name>.sock)` URI
for the service's own endpoint.

Lastly, the aziot-key-openssl-engine-shared engine no longer uses
the `AZIOT_KEY_SERVICE_ENDPOINT` env var, and expects to connect to
`/run/aziot/keyd.sock`