Ensure aziot-keys is built before aziot-keyd. (#39)
aziot-keyd requires aziot-keys to have been built first so that the linker
can link it in. But before this change this was not guaranteed to happen.

This went unnoticed until now because the parallel build would usually finish
building aziot-keys before it got around to linking aziot-keyd.
But Micah happened to have a machine that ended up linking aziot-keyd first
and complained about having no `libaziot_keys.so` to link to.