Disallow custom endpoints and TCP servers in release builds. (#41)
We want to ensure production devices use unix sockets for the transport
in order to have functional authentication of the caller. Furthermore,
in order for host modules to be able to reach the services, the endpoints need
to be in fixed locations. Otherwise every host module would need to provide
its own configurations that the user would have to keep in sync with
any customization they made to the services' endpoints. (Unlike with iotedged,
there is no env var that could give this information to an arbitrary
host module, because host modules run independently of our services.)

To that end, the endpoints configuration for all services will be ignored
in release builds, and the hard-coded defaults of unix sockets will be used.
Sockets inherited via systemd socket activation will also be checked to be
unix sockets.

Custom endpoints can still be configured for debug builds, so local development
and testing can still use unix sockets in custom locations, or TCP sockets.

The sockets directory has been standardized to be `/run/aziot`, rather than
the mix of `/var/lib/aziot` on distros that did not use
systemd socket activation (CentOS 7) and `/var/run/aziot` on distros that did
(everything else). All our supported distros already have `/run`,
so there's no reason to continue using `/var/run`.

Next, this commit enables systemd socket activation on CentOS 7. The original
decision to disable socket activation on CentOS 7 was inherited from iotedged,
which does it because CentOS 7 doesn't support `LISTEN_FDNAMES` and thus
cannot be used for socket-activated services that want more than one socket.
This is the case for iotedged with its mgmt and workload sockets, but is not
the case for the services in this repo. With this commit, socket activation
only uses the `LISTEN_FDS` env var now, and uses the first fd as the socket
that the service should bind to.

Because the endpoints are not configurable, this means the servers can
no longer use *either* systemd socket activation *or* bind a socket themselves.
They must do the former and *fall back* to the latter if the systemd env vars
haven't been set. This is how socket-activated services should work anyway,
so this only makes our services consistent with other software.
Therefore this commit removes the `fd://` URI scheme entirely, and
the default configuration value is the `unix://` URI. If there is
no inherited socket as described above, then the service falls back to
the endpoint URI in the configuration.

Thus, putting all these changes together, the default endpoint URIs are
`unix:///run/aziot/<service_name>.sock` on all distros.

Lastly, the aziot-key-openssl-engine-shared engine no longer uses
the `AZIOT_KEY_SERVICE_ENDPOINT` env var, and expects to connect to
`/run/aziot/keyd.sock`