Use rustfmt (#52)
This PR adds `rustfmt` to the repo, and enforces formatting as part of `make test`.

It's a large diff, but is comprised entirely of formatting changes.

Given the scale of the change (and the fact that it'll likely break all existing un-merged PRs), I'd ask everyone to explicitly approve it before it gets merged. The sooner the better.

Once everyone has given the greenlight, I'll rebase on-top of the latest `main`, and I'll merge it in.