Add basic E2E test workflows.
This commit adds two GH actions that run an E2E test script against
all supported OSes in dynamically-created Azure VMs against
dynamically-created IoT Hubs. The test installs the latest
aziot-identity-service package, configures the provisioning method according to
the test parameters, registers a single host process module, and asserts that
identityd is able to complete provisioning and the module is able to get
its identity.

One action is a scheduled workflow that runs once a day and is only usable on
the Azure/iot-identity-service repo. The other can be triggered manually
via the github.com UI and is primarily meant for forks.

The test script can also be run locally instead of in a GH action.

The script uses an OS-appropriate package from the latest successful
packages workflow run for the specified git branch. When running locally,
it can be told to instead use a locally-built package.

Currently the setup has the following limitations:

- Only one provisioning method is tested - manual+symmetric-key device identity.
  Manual+X509 and DPS tests will be added later.

- Only the amd64 architecture is tested. Testing other architectures
  cannot be done in Azure VMs and will require our own hosted GH agents on
  arm32v7 and aarch64 hardware.

- PKCS#11 is not tested. PKCS#11 tests using tpm2-pkcs11 with ibmswtpm2 can
  be run on the Azure VMs, and will be done in the future.

Other changes:

- Fix file extension check when merging configs.

- Clarify service docs that `/run/aziot` also needs to exist and be writable.

- Use systemd helpers in RPM spec to handle enabling and restarting.