Add API to issue local identities (#61)
`GET /identities/modules/{module-name}?api-version=2020-09-01&type=local`

The query parameter `type=local` causes IS to return a local identity. When not specified, the module identity (if any) will be returned.

Local identities are issued with ephemeral keys not stored by IS. Therefore, every request to IS for a local identity will provide a new cert and key pair.