Fix for SIGPIPE errors (#65)
Without the trailing slash, the server was treating the trailing whitespace
as the start of another request, and thus attempting to respond to it with
an HTTP 400 response. But the sync keyd client had already closed its end of
the connection after reading the first response, which would cause a SIGPIPE
on the server.