For aziot-key-openssl-engine, don't use freef to free the engine's ex data.
A quirk of openssl's dynamic engine loader appears to be that it registers its
own ex data to unload the DSO of the engine it loaded,
ie aziot-key-openssl-engine. However it does this before any other ex data
the engine might have are freed. This means that, before this change,
openssl would try to use aziot_key_freef_engine_ex_data after having already
unloaded the DSO that contained it, leading to a segfault or other behavior.

With this change, the engine's ex data is instead deallocated via
the ENGINE_destroy hook.