Use new rust-toolchain format which removes the need to install components manually.
Also remove the checks around bindgen and cbindgen existing before
`cargo install`ing them. `cargo install` already acts as a no-op if the crate
is installed with the requested version. It even correctly reinstalls the crate
if the currently installed version doesn't match the requested version (both
for upgrading and downgrading), which the previous code did not handle.