add HostConnectDpsEndpoint check
wow there was _very_ little documentation on how to perform a TLS
handshake with hyper_openssl. Turns out you have to treat the
HttpsConnector as a hyper::Service, and then `.call()` it, thereby
getting back a handshake future. There are _zero_ docs which describe
this, and the only reason I figured it out was by crossreferencing the
hyper_openssl implementation with hyper_tls.