Implement support for PKCS#11-backed symmetric keys. (#93)
Since PKCS#11-backed symmetric keys cannot be used for both AES and HMAC,
this change adds a new parameter to the create / import symmetric key API
(at both the libaziot-keys layer and the HTTP API layer) to specify
the key usage.

This change also removes the key length parameter for generating new keys.
Encryption keys may need to be used with AES-256-GCM or AES-128-GCM
depending on what the underlying PKCS#11 library supports, so the key length
needs to be in libaziot-keys' control, not the caller's.

Other changes:

- Delete IS-testing code from iotedged test binary since it no longer works,
  because IS no longer listens on HTTP.