Abort IS if a runtime config update fails (#118)
Identity Service config updates may require reprovisioning with IoT Hub and updating Hub identities. If reprovisioning fails, IS cannot fall back on the old config because the old config may no longer be valid (e.g. the user created a new IoT Hub and deleted the old one). So, IS will abort and the user will need to fix the config and restart IS.