Remove `aziotctl config wizard`
We determined such a prompt-based wizard is too hard to use. For now,
we're shipping the super-config approach used with `aziotctl config apply`,
and are looking to add more focused wizards like `aziotctl config mp` and
`aziotctl config dps` for manual and DPS provisioning respectively
in the future.

As part of removing the mentions of `aziotctl config wizard`
from `packaging.md`, this commit also removes the IoT Edge-specific text from
that document. This text will be moved to a matching document in
the iotedge repo and be maintained there.

Co-authored-by: Micah Lewis <1009123+micahl@users.noreply.github.com>