Remove hyper dependency in aziotctl-common (#170)
Building iotedge after bumping iot-identity-service dependencies resulted in error due to hyper getting pulled in with after Azure/iot-identity-service#163. Relocated reprovision to become a private method in `aziotctl` which is the only caller.

Dependency issue being fixed:
```
error: failed to select a version for `hyper`.
    ... required by package `http-common v0.1.0 (https://github.com/Azure/iot-identity-service?branch=main#ccdb743a)`
    ... which is depended on by `aziot-certd-config v0.1.0 (https://github.com/Azure/iot-identity-service?branch=main#ccdb743a)`
    ... which is depended on by `aziotctl-common v0.1.0 (https://github.com/Azure/iot-identity-service?branch=main#ccdb743a)`
    ... which is depended on by `iotedge v0.1.0 (/s/iotedge/edgelet/iotedge)`
versions that meet the requirements `^0.14` are: 0.14.4, 0.14.3, 0.14.2, 0.14.1, 0.14.0
```