Remove override of tpm2-pkcs11 store file. (#219)
The existing instructions for creating the base slot were wrong because
`sudo -u aziotks ./tpm2_ptool` would not see the `TPM2_PKCS11_STORE` env var.
But we don't really care to override the default store location anyway;
the default will put it in `$HOME/.tpm2_pkcs11` where `$HOME` is
aziotks's homedir, which is just as good a place for it as anywhere else.

If the user does want to override the store location for some reason,
they can do it on their own.