Allow shutting down http_common server (#249)
Passes in a shutdown oneshot to `http_common::Incoming:serve`

Posting a message to the corresponding `oneshot::Sender` will cause the server to gracefully shut down by awaiting all unfinished server tasks.

This is currently not used by any of the aziotd services, but edged will use it for when it has to shut down and restart its servers for reprovisioning.