Add headers option to request_no_content (#260)
This pr adds a headers function for request_no_content. It also slightly changes the behavior or request_no_content, allowing responses other than 204 to return Ok. This is because sometimes when docker-rs calls the no_content function, the http response has content, we just don't care about the content and want to drop it. 

This pr also pulls the shared code of both request functions in to shared functions to reduce copied code and make them easier to read.