Emit base64-encoded DER in 64-character chunks (#323)
Cisco's libest implementation[^1] does not accept non-chunked inputs, and pull
296 inadvertently broke support for it by switching to an unchunked base64
encoding[^2].  Though libest is not advertised as an EST server suitable for
production, it is still popular enough to warrant accommodation by our client.
Since it seems that EST servers that accept unchunked input also accept chunked
input, it should be safe (with respect to compatibility with other servers) to
make this change.

[^1]: https://github.com/cisco/libest
[^2]: https://github.com/Azure/iot-identity-service/commit/550dad87a3222858137395343c82c67fc14e6fdd