Ensure parent directories are readable by target user when creating directories in `config apply`
Cherry-pick from main of 299014df6652a5491fe702a5236b35140119c2c2

When `config apply` creates the /var/secrets/aziot/keyd directory to store
the inline symmetric key, it assumed that parent directories would already
be readable by the target user. But this is not a valid assumption since
it depends on root's umask, and we've had a few cases of devices where
root's umask is such that new directories are 0700 or 0750 by default.

This change adds a step of walking through the parents,
checking if they're readable by the target user, and making them
world-readable if they're not.