Make apt upgrade non-interactive for VM provisioning (#373)
This PR fixes persistent failures in ubuntu:18.04 jobs.

Currently, the VM provisioning process is unable to run **apt upgrade**, which in turn is because the upgrade of the openssh-server package is conflicting with files that are pre-installed on the VM image.

This change makes the upgrades non-interactive and automatically chooses the newer file(s) in case of any conflict (which is fine since it's a clean VM).


**Testing Notes:**

- Check-in gates[ passed (on my fork)](https://github.com/gauravIoTEdge/iot-identity-service/actions/runs/1890462264)
- Ran [E2E tests (on my fork)](https://github.com/gauravIoTEdge/iot-identity-service/runs/5313085676?check_suite_focus=true) and a couple of 18.04 jobs passed, unlike before.