Create Mariner 2.0 Build for Iot-Identity-Service (#377)
Adds an additional build for mariner 2.0. This requires a separate build of the mariner toolkit to be used. As older packages are not being backported to mariner 2.0, this means we would not have access to the correct versions of GCC and Rust. A patch to the mariner spec was added file which prevents warnings from c-shared being propagated as errors when building. The patch also modifies the arguments to the hmacResult function as they were mismatched causing a warning which would result in an error. These warnings are a result of using a gcc version greater that 9 and Mariner 2.0 only provides 11+. 

This also will utilize the rust toolkit provided by the install-build-deps.sh script. this will then be packaged and forwarded to the spec files sources directory.