Add move_key_pair API (#382)
Adds an API to move a key pair from one key ID to another. This has the effect of replacing the destination key ID with the source and deleting the source key ID.

This API will be used to support rotation of keys managed by Keys Service during cert renewal.