[main] Upgrade to latest Rust version (#390)
This is our monthly toolchain upgrade to the latest stable version. Since the primary artifacts of this repository are binaries, we want to be tracking Rust compiler releases closely in the event a stdlib vulnerability is found. We do not pin to "stable", however, since that has caused pipeline breakage when some code patterns are made illegal (like in the upgrade from 1.47 to 1.48 [^0]) or, more often, clippy lints are added.

- Use `std::ptr::addr_of` and `std::ptr::addr_of_mut` for pointers to places
- Use `NonNull` in `pkcs11` shared library code
- Switch from `TryInto::try_into` to `.try_into` since we are now on Rust 2021
- Remove redundant `try_into` calls

[^0]: https://github.com/rust-lang/rust/blob/master/RELEASES.md#compatibility-notes-11
  Namely, the point on `mem::uninitialized`.