[Cherry-Pick] Add aarch64 Build for Mariner 1.0 and 2.0 to main branch (#412)
Cherry Pick from: https://github.com/Azure/iot-identity-service/pull/408

Adds aarch64 build for Mariner 1.0 and 2.0. This will not build the aarch64 version in the current package build pipelines. this will instead update the corresponding build scripts to work for aarch64 machines. The mariner aarch64 builds were not added to the normal package build yaml since mariner does not support cross-compiling and requires the packages to be built in a native aarch64 machine.