[aziot-tpmd] Replace uTPM with libtss (#415)
`aziot-tpmd` currently uses uTPM[^1] for interacting with the host TPM. However, it is no longer actively maintained, and we would also like to resolve some of the limitations of our `tpm/aziot-tpm-sys/azure-iot-hsm-c` library. To this end, we replace uTPM with the tpm2-tss library, an open-source implementation of the TCG TPM2 Software Stack[^2].

Immediate limitations lifted by this PR are support of TPM hierarchy authorization values and DPS attestation key index immovability. By leveraging the tss2-tctildr library, we are also able to accommodate more TPM configurations---including simulated TPMs---with only a configuration change to the TCTI loader string.

[^1]: https://github.com/Azure/azure-utpm-c
[^2]: https://github.com/tpm2-software/tpm2-tss