Support DPS custom allocation payloads (#430)
Adds the ability to specify a custom JSON payload in config.toml. This payload is sent to DPS during device registration. This feature allows users to configure DPS with a custom Azure Functions webhook that reads the custom payload and uses it to determine which IoT hub to assign the device to. See the following Microsoft Docs page for more context: https://docs.microsoft.com/en-us/azure/iot-dps/how-to-send-additional-data.