Add no-socket-throttle feature to http-common (#438)
Adds a cargo feature `no-socket-throttle` to http-common. Enabling this disables socket throttling when serving connections on UNIX sockets.

This is in response to #433. The 1.3.0 update consolidated the server code between iotedge and this repo, and unintentionally introduced this repo's throttling mechanism to the Edge daemon. The Edge daemon does not need to throttle the management socket, as only users in the iotedge group can access the management socket. By default, this prevents modules from calling the management socket.