Upgrade to Rust 1.64 (#462)
This is our monthly toolchain upgrade to the latest stable version. Since the primary artifacts of this repository are binaries, we want to be tracking Rust compiler releases closely in the event a stdlib vulnerability is found. We do not pin to "stable", however, since that has caused pipeline breakage when some code patterns are made illegal (like in the upgrade from 1.47 to 1.48 [^0]) or, more often, clippy lints are added.

- `clippy::unnecessary_lazy_evaluations`: replace `${bool}.then(|| ${value})` with `${bool}.then_some(${value})`
- `clippy::unused_async`: remove `async` where functions are fully implemented and unlikely to change, otherwise allow with inline note explaining why lint is allowed
- `clippy::unused_self`: change methods to associated functions

[^0]: https://github.com/rust-lang/rust/blob/1.48.0/RELEASES.md#compatibility-notes
  Namely, the point on `mem::uninitialized`.