Add Ubuntu 22.04 as a target (#463)
The existing openssl engine is used with openssl 3.0, leveraging the existing APIs that are marked as deprecated and expected to be removed after 3.0.

This change also moves the cargo build out of the .deb package process to occur prior to running dpkg-buildpackage.  Otherwise, the build hits an error where the linker complains about duplicate symbols for `bind_engine` and `v_check`.  That error only happens on 22.04 and only when the `default` makefile target is done as part of dpkg-buildpackage.