Upgrade to Rust 1.65 (#484)
This is our monthly toolchain upgrade to the latest stable version. Since the primary artifacts of this repository are binaries, we want to be tracking Rust compiler releases closely in the event a stdlib vulnerability is found. We do not pin to "stable", however, since that has caused pipeline breakage when some code patterns are made illegal (like in the upgrade from 1.47 to 1.48 [^0]) or, more often, clippy lints are added.

- `clippy::bool_to_int_with_if`: replace `if ${BOOL} { 1 } else { 0 }` with `${BOOL}.into()`
- `clippy::explicit_auto_deref`: remove redundant dereferences
- `clippy::manual_string_new`: replace `"".to_owned()` with `String::new()`
- `clippy::needless_borrow`: remove unnecessary borrows

Switching to `let .. else` where appropriate was skipped in the interest of keeping the changeset small.

[^0]: https://github.com/rust-lang/rust/blob/1.48.0/RELEASES.md#compatibility-notes
  Namely, the point on `mem::uninitialized`.