Use plaintext buffer equal in length to the ciphertext buffer for PKCS#11 decrypt. (#487)
The previous code used a plaintext buffer that was 16 bytes smaller than
the ciphertext buffer, because the ciphertext was known to include
a 16-byte tag. However at least SoftHSM requires the plaintext buffer to be
at least as big as the ciphertext buffer. So this commit makes it so.

Fixes #486