Disable X.509 E2E tests on Ubuntu 22.04 (#516)
A bug in openssl 3, which is used in Ubuntu 22.04, causes an error when attempting to use openssl s_client with an engine. For now, disable the X.509 Ubuntu 22.04 E2E test runs so that the E2E tests can pass.

The openssl bug is tracked here: https://github.com/openssl/openssl/issues/20161
Revert this commit once the bug is fixed.