Do not require CAP_FOWNER for aziotctl_common::config::write_file
The aziotctl_common::config::write_file function first changes ownership
of a file, then changes the permissions.

Changing the permissions on a file owned by another user requires
CAP_FOWNER. While this is not normally a problem for root, the
assumption that root ALWAYS has CAP_FOWNER is erroneous (snaps are a
counter example).

This commit simply reorders the operation, so that the file is still
owned by the creator when the permissions get changed.

Signed-off-by: Alex Lewontin <alex.lewontin@canonical.com>