Disable X.509 E2E tests on RHEL 9 (#524)
Applies changes from #516 to RHEL 9, which also uses openssl3:

A bug in openssl 3, which is used in Ubuntu 22.04 [and RHEL 9], causes an error when attempting to use openssl s_client with an engine. For now, disable the X.509 Ubuntu 22.04 [and RHEL9] E2E tests runs so that the E2E tests can pass.

This change should be reverted when the openssl bug is fixed.