When creating an AES PKCS#11 key, test it with a dummy encryption to ensure that the token can perform AES-GCM.
It is possible for PKCS#11 implementations to support AES keys only with
AES-CBC etc but not with AES-GCM. Before this change, creating the key would
succeed, and then encrypting with it would fail, at which point it would be
too late.

With this change, we do a dummy encruption with the key just to be sure that
it can be used with AES-GCM. If that fails, we continue to the next location,
usually the filesystem ie openssl keys.