Set snap grade according to version (#596)
Snaps are currently built with grade=devel, which means they can't be promoted to the stable channel. We want dev builds (e.g., builds from main) to keep grade=devel so they can't accidentally be promoted, but we want release builds to be grade=stable. This change adds the logic to do that, and is patterned after similar logic in the azure-iot-edge snap.