Add back CentOS 7 & Debian 10 for 1.4 end-to-end tests (#607)
Support for CentOS 7 and Debian 10 were recently removed from the end-to-end test workflows in main because they will not be supported in version 1.5. But the workflow in main runs the tests for both main and release/1.4 branches and those platforms are still supported in 1.4 until they reach end of life (June 2024). This change adds them back for 1.4.